[close #357] Actual projection is now used for cropping polygons 1 meters outside the frame when exporting

+++ src/core/tmexportdatashp.h
@@ -53,7 +53,7 @@ class tmExportDataSHP : public tmExportData
 		OGRGeometry * SafeIntersection(OGRGeometry * line, OGRGeometry * frame);
 		OGRGeometry * SafeUnion (OGRGeometry * union1, OGRGeometry * line);
 		OGRGeometry * SafeCreateFromGEOS (GEOSGeom geosGeom);
-		OGRGeometry * SafeBuffer (OGRGeometry * ogrgeom,  int size);
+		OGRGeometry * SafeBuffer (OGRGeometry * ogrgeom,  double size);
 		
 		// multiple fields
 		bool SetMultipleFields (ProjectDefMemoryLayers * layer,
